test/requirements.txt
pyserial
python-escpos
pytest
